+++ History.md
@@ -1,6 +1,13 @@
 Dalli Changelog
 =====================
 
+HEAD
+-----
+
+ - Major design change - raw support is back to maximize compatibility with Rails
+ and the increment/decrement operations.  You can now pass :raw => true to most methods
+ to bypass (un)marshalling.
+
 0.9.4
 -----